// apps/pilote-ppg/src/client/components/_commons/InputGroupe/InputGroupe.interface.ts
export default interface InputGroupeProps {
  optionsGroupées: InputGroupeOptionsGroupées;
  valeurSelectionneeParDefaut: string;
  changementValeurSelectionneeCallback: (valeursSélectionnées: string) => void;
  label: string;
  libelle: string;
  desactive?: boolean;
  direction: "horizontal" | "vertical";
}

export type InputGroupeOption = {
  label: string;
  value: string;
  disabled?: boolean;
  afficherIcone?: boolean;
};

export type InputGroupeOptionGroupée = {
  label: string;
  options: InputGroupeOptions;
};

export type InputGroupeOptions = InputGroupeOption[];
export type InputGroupeOptionsGroupées = InputGroupeOptionGroupée[];